Working on protection against self removal by admin

+++ src/main/java/pl/coderslab/charity/controller/AdminController.java
@@ -159,13 +159,15 @@ public class AdminController {
     }
 
     @GetMapping("/delete/{id}")
-    public String delete(@PathVariable Long id){
+    public String delete(@PathVariable Long id, @AuthenticationPrincipal CurrentUser customUser){
         Optional<User> optionalUser = userService.get(id);
         User admin = optionalUser.orElse(null);
         if(admin == null){
             return "problemAdmin";
         }
-
+        if(admin.getId().equals(customUser.getUser().getId())){
+            return "deleteNotPossible";
+        }
         userService.delete(id);
         return "redirect:/admins/list";
     }
